requirements/base.txt
Django==5.0.4
bpython==0.24
celery[redis]==5.3.6
django-cacheops==7.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.15.1
drf-spectacular==0.27.2
pytz==2024.1
